backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
mysqlclient==2.2.4
sqlparse==0.5.0
